费用分摊修改分摊值开发

+++ erp-model/erp-model-tms/src/main/java/com/erp/model/tms/dto/FirstMileWeightAllocationDTO.java
@@ -938,17 +938,17 @@ public class FirstMileWeightAllocationDTO implements Serializable {
         /**
          * 箱长
          */
-        private Integer boxLength;
+        private BigDecimal boxLength;
 
         /**
          * 箱宽
          */
-        private Integer boxWidth;
+        private BigDecimal boxWidth;
 
         /**
          * 箱高
          */
-        private Integer boxHeight;
+        private BigDecimal boxHeight;
 
         /**
          * 箱子尺寸单位
@@ -1014,19 +1014,19 @@ public class FirstMileWeightAllocationDTO implements Serializable {
          * 箱长
          */
         @NotNull(message = "新的箱长不能为空")
-        private Integer newBoxLength;
+        private BigDecimal newBoxLength;
 
         /**
          * 箱宽
          */
         @NotNull(message = "新的箱宽不能为空")
-        private Integer newBoxWidth;
+        private BigDecimal newBoxWidth;
 
         /**
          * 箱高
          */
         @NotNull(message = "新的箱高不能为空")
-        private Integer newBoxHeight;
+        private BigDecimal newBoxHeight;
 
         /**
          * 出库重量
